+++ src/renderer/src/components/message/UserMessageBlock.tsx
@@ -1,5 +1,6 @@
 import { CgTerminal } from 'react-icons/cg';
 import { FaRegQuestionCircle } from 'react-icons/fa';
+import { AiOutlineFileSearch } from 'react-icons/ai';
 import { IoConstruct } from 'react-icons/io5';
 
 import { CopyMessageButton } from './CopyMessageButton';
@@ -18,10 +19,11 @@ export const UserMessageBlock = ({ baseDir, message, allFiles }: Props) => {
 
   return (
     <div className={`${baseClasses} relative flex items-start gap-3 group`}>
-      <div className="flex items-center">
-        {(message.editFormat === 'code' || !message.editFormat) && <CgTerminal className="text-neutral-200 h-[18px]" />}
-        {message.editFormat === 'ask' && <FaRegQuestionCircle className="text-neutral-200 h-[18px]" />}
-        {message.editFormat === 'architect' && <IoConstruct className="text-neutral-200 h-[18px]" />}
+      <div className="mt-[1px]">
+        {(message.editFormat === 'code' || !message.editFormat) && <CgTerminal className="text-neutral-200" size={16} />}
+        {message.editFormat === 'ask' && <FaRegQuestionCircle className="text-neutral-200" size={16} />}
+        {message.editFormat === 'architect' && <IoConstruct className="text-neutral-200" size={16} />}
+        {message.editFormat === 'context' && <AiOutlineFileSearch className="text-neutral-200" size={16} />}
       </div>
       <div className="flex-grow-1">{parseMessageContent(baseDir, message.content, allFiles)}</div>
       <div className="absolute top-2 right-2">